% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batis_download.R
\name{batis_download}
\alias{batis_download}
\title{OECD BaTIS API Downloader}
\usage{
batis_download(
  reporter_code,
  start_year,
  end_year = start_year,
  output_dir = ".",
  key_from_site,
  format = "csvfilewithlabels"
)
}
\arguments{
\item{reporter_code}{3-letter ISO code (e.g., "FRA")}

\item{start_year}{Integer start year}

\item{end_year}{Integer end year (defaults to start_year)}

\item{output_dir}{Output folder for the xlsx}

\item{key_from_site}{Long key copied from the OECD site}

\item{format}{"csvfilewithlabels" (default) or "csvfile"}
}
\value{
Invisibly, the output file path
}
\description{
Download OECD BaTIS and write Excel, excludes self-as-partner,
and saves as "[REPORTER]_BATIS_[start]_[end].xlsx".
}
\examples{
\dontrun{
key <- "AUS.AUT+BEL+CAN+CHL+COL+CRI+CZE+DNK+EST+FIN+FRA+DEU+GRC+HUN+ISL+IRL+ISR+ITA+JPN+KOR+LVA+LTU+LUX+MEX+NLD+NZL+NOR+POL+PRT+SVK+SVN+ESP+SWE+CHE+TUR+GBR+USA+AFG+ALB+DZA+AGO+AIA+ATG+ARG+ARM+ABW+AZE+BHS+BHR+BGD+BRB+BLR+BLZ+BEN+BMU+BTN+BOL+BIH+BWA+BRA+BRN+BGR+BFA+BDI+CPV+KHM+CMR+CYM+CAF+TCD+CHN+COM+COG+CIV+HRV+CUB+CUW+CYP+PRK+COD+DJI+DMA+DOM+ECU+EGY+SLV+GNQ+ERI+SWZ+ETH+FRO+FJI+PYF+GAB+GMB+GEO+GHA+GRD+GTM+GIN+GNB+GUY+HTI+HND+HKG+IND+IDN+IRN+IRQ+JAM+JOR+KAZ+KEN+KIR+XKV+KWT+KGZ+LAO+LBN+LSO+LBR+LBY+MAC+MDG+MWI+MYS+MDV+MLI+MLT+MRT+MUS+MDA+MNG+MNE+MSR+MAR+MOZ+MMR+NAM+NPL+NCL+NIC+NER+NGA+MKD+OMN+PAK+PSE+PAN+PNG+PRY+PER+PHL+QAT+ROU+RUS+RWA+KNA+LCA+VCT+WSM+STP+SAU+SEN+SRB+SYC+SLE+SGP+SXM+SLB+SOM+ZAF+LKA+SDN+SUR+SYR+TWN+TJK+TZA+THA+TLS+TGO+TON+TTO+TUN+TKM+TCA+TUV+UGA+UKR+ARE+URY+UZB+VUT+VEN+VNM+YEM+ZMB+ZWE+ANT_F+SCG_F+W+WXD.M+X..SC1+SDA+SI1+SJ1+SK1+SC2+SDB+SI2+SJ2+SK2+SC3+SI3+SJ3+SC4+SA+SB+SC+SD+SE+SF+SG+SH+SI+SJ+SK+SL+SOX+SOX1+SPX1+SPX4+S.A.USD_EXC.B"
batis_download(
reporter_code = "JPN",         #CHANGE COUNTRY CODE
start_year    = 2018,          #CHANGE
end_year      = 2018,          #CHANGE
output_dir    = '/Users/cmbzamora/Library/CloudStorage/OneDrive-AsianDevelopmentBank/Personal/MRIO Series Revisions 2017-24/JPN/BATIS',   #PATH / DIRECTORY WHERE YOU WANT THE FILE TO BE SAVED
key_from_site = key            #<- DO NOT CHANGE
)
}
}
